Alterações da regra de negócio

+++ TDE/src/main/java/com/example/tde/repository/CategoriaRepository.java
@@ -1,5 +1,6 @@
 package com.example.tde.repository;
 
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -9,6 +10,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CategoriaRepository extends JpaRepository <Categoria, Long>{
-    
+public interface CategoriaRepository extends JpaRepository<Categoria,Integer> {
+
 }
+
