+++ TDE/src/main/java/com/example/tde/repository/CidadeRepository.java
@@ -4,11 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.example.tde.domain.Cidade;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
 
 @Repository
-public interface CidadeRepository extends JpaRepository <Cidade, Long> {
+public interface CidadeRepository extends JpaRepository <Cidade, Integer> {
     
 }
